feat: move getting started content into a category provider
